requirements.txt
numpy
tensorflow>=1.12.1
opencv-python
opencv-contrib-python
keras
matplotlib
pillow
jupyter
pandas
imageio
scikit-image
